Restructure source code
