+++ externs/src/ipfix/export.cpp
@@ -1,0 +1,9 @@
+#include "ipfix.h"
+
+void export_flow_records(FlowRecordCache_t &records) {
+  for (auto i = records.begin(); i != records.end(); ++i) {
+    auto record = i->second;
+    std::cout << "IPFIX EXPORT: Exporting record:" << std::endl;
+    std::cout << record << std::endl;
+  }
+}
